Started on rendering the attendance at locations
Added logic to group data using the d3.nest() functions, where data is grouped by the year, and the attendance numbers are aggregated. Interestingly, the locations are also aggregated by getting the mean of the locations where the games occurred during that year.

This location mean covers the scenario for when a world cup occurred in multiple locations. Since we want to just print one dot to give a general idea of the location on the globe where the tournament occurred, we take the mean of the latitude and longitude values to get the mean location (between all those points).