Sample component with React and D3
Playing with D3 and React. Initial commit. The basic theme appears to be to let React handle the rendering of the data points but allow D3 access to the DOM to set the axes. I need to play a little more to see whether we can use D3's selectall() method to place points.