Added the axes and the legend
Interesting how the legend is drawn. I deviated from the example by positioning the whole legend as a group to one side of the page, and then positioned the individual text and rects for the legend within that group. This way, if I want to make changes to the placement of the legend, I only have the change the x setting at the entire legend group level and not at each line of the legend.