Stacked Area Chart, showing up the way I want it :)
The trick here is in setting the max y value. To get the charts to stack the way you want, you need to set the max y value to the *sum* of all the y values. To get it to render full screen as a percentage of a whole, you need to convert the y data to a fractional percentage value.

Signed-off-by: nikhilpanchal <nikhilpanchal@gmail.com>