Added a header and css to go with it
Got the header from the google developers website, as I liked the shadow effect. Interesting exercise looking at their css values and then applying to this site. The idea is to create this header here and then use it as a reference point to create the same for any other website using the same values.

Signed-off-by: nikhilpanchal <nikhilpanchal@gmail.com>